[CMake] Rename WinCairo port to Win https://bugs.webkit.org/show_bug.cgi?id=251927
Reviewed by Ross Kirsling.

Apple Windows port has gone. Rename WinCairo to Win. All scripts
should accept both --win and --wincairo switches for the transition
period.

* Source/WTF/WTF.xcodeproj/project.pbxproj:
* Source/WTF/wtf/CMakeLists.txt:
* Source/WTF/wtf/PlatformEnable.h:
* Source/WTF/wtf/PlatformEnableWin.h: Renamed from Source/WTF/wtf/PlatformEnableWinCairo.h.
* Source/WebCore/platform/graphics/gstreamer/MediaPlayerPrivateGStreamer.h:
* Source/cmake/OptionsWin.cmake:
* Source/cmake/OptionsWinCairo.cmake: Removed.
* Source/cmake/WebKitCommon.cmake:
* Tools/Scripts/build-webkit:
* Tools/Scripts/webkitdirs.pm:
(argumentsForConfiguration):
(extractNonMacOSHostConfiguration):
(determineConfigurationProductDir):
(determinePortName):
(isWin):
(setPathForRunningWebKitApp):
* Tools/Scripts/webkitperl/BuildSubproject.pm:
* Tools/Scripts/webkitperl/test262_unittest/test262-runner-tests.pl:
* Tools/Scripts/webkitperl/webkitdirs_unittest/extractNonMacOSHostConfiguration.pl:

Canonical link: https://commits.webkit.org/280891@main
