--- Source/cmake/OptionsWinCairo.cmake
@@ -1,2 +1,0 @@
-# FIXME: https://bugs.webkit.org/show_bug.cgi?id=251927
-include(OptionsWin)
