+++ Tools/Scripts/build-webkit
@@ -132,7 +132,7 @@ Usage: $programName [options] [options to pass to build system]
 
   --gtk                             Build the GTK+ port
   --wpe                             Build the WPE port
-  --wincairo                        Build the Windows port
+  --win                             Build the Windows port
   --playstation                     Build the PlayStation port
 
   --inspector-frontend              Copy Web Inspector user interface resources to the build directory
@@ -211,7 +211,7 @@ push @projects, ("Source/WebKitLegacy");
 
 
 # Build WTF as a separate static library on ports which support it.
-splice @projects, 0, 0, "Source/WTF" if isWinCairo();
+splice @projects, 0, 0, "Source/WTF" if isWin();
 
 splice @projects, 0, 0, "Source/bmalloc" if isAppleCocoaWebKit();
 
@@ -224,7 +224,7 @@ if (!isCMakeBuild()) {
     }
 }
 
-if ((isAppleWebKit() || isWinCairo() || isPlayStation()) && !-d "WebKitLibraries") {
+if ((isAppleWebKit() || isWin() || isPlayStation()) && !-d "WebKitLibraries") {
     die "Error: No WebKitLibraries directory found. Please do a fresh checkout.\n";
 }
 
@@ -288,7 +288,7 @@ if (isAppleCocoaWebKit()) {
         push @projects, ("Tools/DumpRenderTree", "Tools/WebKitTestRunner", "Source/ThirdParty/gtest", "Tools/TestWebKitAPI");
     }
 
-} elsif (isWinCairo() && !$skipLibraryUpdate) {
+} elsif (isWin() && !$skipLibraryUpdate) {
     (system("python Tools/Scripts/update-webkit-wincairo-libs.py") == 0) or die;
 }
 
@@ -324,7 +324,7 @@ if (isCMakeBuild() && !isAnyWindows()) {
 }
 
 my $baseProductDir = baseProductDir();
-if (isWinCairo() || isPlayStation()) {
+if (isWin() || isPlayStation()) {
     my @featureArgs = cmakeArgsFromFeatures(@features, !$noExperimentalFeatures);
     removeCMakeCache(@featureArgs);
