+++ Tools/Scripts/webkitperl/BuildSubproject.pm
@@ -135,7 +135,7 @@ if ($buildDir && !isCMakeBuild()) {
     exit 1;
 }
 
-if (isWinCairo() || (isJSCOnly() && isWindows())) {
+if (isWin() || (isJSCOnly() && isWindows())) {
     (system("python3 Tools/Scripts/update-webkit-wincairo-libs.py") == 0) or die;
 }
